+++ src/CPG.Application/UseCases/Companies/ViewModels/CreateCompanyModel.cs
@@ -17,7 +17,7 @@ public class CreateCompanyModel
     [Required]
     public short IpgRedirectionMethodType { get; set; }
     [Required]
-    public required short[] MethodTypes { get; set; }
+    public required byte[] MethodTypes { get; set; }
     [Required]
     public required List<long> Users { get; set; }
     [Required]
